requirements.txt
azure-storage-blob==12.13
flask==2.1.3